use truecolor in colorer by adopting truemod functionality and extending editor API

+++ far2l/src/console/scrbuf.hpp
@@ -81,10 +81,10 @@ class ScreenBuf
 		void RestoreMacroChar();
 		void RestoreElevationChar();
 
-		void ApplyColorMask(int X1,int Y1,int X2,int Y2,WORD ColorMask);
-		void ApplyColor(int X1,int Y1,int X2,int Y2,WORD Color);
-		void ApplyColor(int X1,int Y1,int X2,int Y2,int Color,WORD ExceptColor);
-		void FillRect(int X1,int Y1,int X2,int Y2,WCHAR Ch,WORD Color);
+		void ApplyColorMask(int X1,int Y1,int X2,int Y2,DWORD64 ColorMask);
+		void ApplyColor(int X1,int Y1,int X2,int Y2,DWORD64 Color);
+		void ApplyColor(int X1,int Y1,int X2,int Y2,DWORD64 Color,DWORD64 ExceptColor);
+		void FillRect(int X1,int Y1,int X2,int Y2,WCHAR Ch,DWORD64 Color);
 
 		void Scroll(int);
 		void Flush();
